bugfix ENR sequence number to 64 bits
See:
https://github.com/ethereum/devp2p/blob/6eddaf50298d551a83bcc242e7ce7024c6cc8590/enr.md

+++ src/alexandria/protocol.rs
@@ -143,7 +143,7 @@ impl AlexandriaProtocol {
             Request::Ping(Ping { .. }) => {
                 let enr_seq = self.discovery.local_enr().seq();
                 Response::Pong(Pong {
-                    enr_seq: enr_seq as u32,
+                    enr_seq: enr_seq,
                     data_radius: self.data_radius,
                 })
             }
